+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 8.7
 - Add ProgramFunctions/SendEmail.fnc.php|send_error action hook in SendEmail.fnc.php
 - EasyCodingStandard use short array notation in functions/, classes/core/, ProgramFunctions/, modules/ & plugins/
 - Fix month + year format, remove day (regression since 7.1) in Dashboard.inc.php
+- ProgramUserConfig() always return array, not null in Config.fnc.php & \_makeLetterGrade.fnc.php
 
 Changes in 8.6.1
 ----------------